[SPARK-46433][PS][TESTS] Reorganize `OpsOnDiffFramesGroupByTests`: Factor out `test_aggregate/apply/cum*`
### What changes were proposed in this pull request?
Factor out `test_aggregate/apply/cum*`

### Why are the changes needed?
they are slow tests

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44385 from zhengruifeng/ps_test_diff_groupby_1.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
