+++ python/pyspark/pandas/tests/test_ops_on_diff_frames_groupby.py
@@ -167,89 +167,6 @@ class OpsOnDiffFramesGroupByTestsMixin:
             pdf1.B.rename().groupby(pdf2.A.rename()).sum().sort_index(),
         )
 
-    def test_aggregate(self):
-        pdf1 = pd.DataFrame({"C": [0.362, 0.227, 1.267, -0.562], "B": [1, 2, 3, 4]})
-        pdf2 = pd.DataFrame({"A": [1, 1, 2, 2]})
-        psdf1 = ps.from_pandas(pdf1)
-        psdf2 = ps.from_pandas(pdf2)
-
-        for as_index in [True, False]:
-            if as_index:
-
-                def sort(df):
-                    return df.sort_index()
-
-            else:
-
-                def sort(df):
-                    return df.sort_values(list(df.columns)).reset_index(drop=True)
-
-            with self.subTest(as_index=as_index):
-                self.assert_eq(
-                    sort(psdf1.groupby(psdf2.A, as_index=as_index).agg("sum")),
-                    sort(pdf1.groupby(pdf2.A, as_index=as_index).agg("sum")),
-                )
-                self.assert_eq(
-                    sort(psdf1.groupby(psdf2.A, as_index=as_index).agg({"B": "min", "C": "sum"})),
-                    sort(pdf1.groupby(pdf2.A, as_index=as_index).agg({"B": "min", "C": "sum"})),
-                )
-                self.assert_eq(
-                    sort(
-                        psdf1.groupby(psdf2.A, as_index=as_index).agg(
-                            {"B": ["min", "max"], "C": "sum"}
-                        )
-                    ),
-                    sort(
-                        pdf1.groupby(pdf2.A, as_index=as_index).agg(
-                            {"B": ["min", "max"], "C": "sum"}
-                        )
-                    ),
-                )
-                self.assert_eq(
-                    sort(psdf1.groupby([psdf1.C, psdf2.A], as_index=as_index).agg("sum")),
-                    sort(pdf1.groupby([pdf1.C, pdf2.A], as_index=as_index).agg("sum")),
-                )
-                self.assert_eq(
-                    sort(psdf1.groupby([psdf1.C + 1, psdf2.A], as_index=as_index).agg("sum")),
-                    sort(pdf1.groupby([pdf1.C + 1, pdf2.A], as_index=as_index).agg("sum")),
-                )
-
-        # multi-index columns
-        columns = pd.MultiIndex.from_tuples([("Y", "C"), ("X", "B")])
-        pdf1.columns = columns
-        psdf1.columns = columns
-
-        columns = pd.MultiIndex.from_tuples([("X", "A")])
-        pdf2.columns = columns
-        psdf2.columns = columns
-
-        for as_index in [True, False]:
-            stats_psdf = psdf1.groupby(psdf2[("X", "A")], as_index=as_index).agg(
-                {("X", "B"): "min", ("Y", "C"): "sum"}
-            )
-            stats_pdf = pdf1.groupby(pdf2[("X", "A")], as_index=as_index).agg(
-                {("X", "B"): "min", ("Y", "C"): "sum"}
-            )
-            self.assert_eq(
-                stats_psdf.sort_values(by=[("X", "B"), ("Y", "C")]).reset_index(drop=True),
-                stats_pdf.sort_values(by=[("X", "B"), ("Y", "C")]).reset_index(drop=True),
-            )
-
-        stats_psdf = psdf1.groupby(psdf2[("X", "A")]).agg(
-            {("X", "B"): ["min", "max"], ("Y", "C"): "sum"}
-        )
-        stats_pdf = pdf1.groupby(pdf2[("X", "A")]).agg(
-            {("X", "B"): ["min", "max"], ("Y", "C"): "sum"}
-        )
-        self.assert_eq(
-            stats_psdf.sort_values(
-                by=[("X", "B", "min"), ("X", "B", "max"), ("Y", "C", "sum")]
-            ).reset_index(drop=True),
-            stats_pdf.sort_values(
-                by=[("X", "B", "min"), ("X", "B", "max"), ("Y", "C", "sum")]
-            ).reset_index(drop=True),
-        )
-
     def test_duplicated_labels(self):
         pdf1 = pd.DataFrame({"A": [3, 2, 1]})
         pdf2 = pd.DataFrame({"A": [1, 2, 3]})
@@ -262,32 +179,6 @@ class OpsOnDiffFramesGroupByTestsMixin:
         self.assert_eq(
             psdf1.groupby(psdf2.A, as_index=False).sum().sort_values("A").reset_index(drop=True),
             pdf1.groupby(pdf2.A, as_index=False).sum().sort_values("A").reset_index(drop=True),
-        )
-
-    def test_apply(self):
-        pdf = pd.DataFrame(
-            {"a": [1, 2, 3, 4, 5, 6], "b": [1, 1, 2, 3, 5, 8], "c": [1, 4, 9, 16, 25, 36]},
-            columns=["a", "b", "c"],
-        )
-        pkey = pd.Series([1, 1, 2, 3, 5, 8])
-        psdf = ps.from_pandas(pdf)
-        kkey = ps.from_pandas(pkey)
-
-        self.assert_eq(
-            psdf.groupby(kkey).apply(lambda x: x + x.min()).sort_index(),
-            pdf.groupby(pkey).apply(lambda x: x + x.min()).sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)["a"].apply(lambda x: x + x.min()).sort_index(),
-            pdf.groupby(pkey)["a"].apply(lambda x: x + x.min()).sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)[["a"]].apply(lambda x: x + x.min()).sort_index(),
-            pdf.groupby(pkey)[["a"]].apply(lambda x: x + x.min()).sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(["a", kkey]).apply(lambda x: x + x.min()).sort_index(),
-            pdf.groupby(["a", pkey]).apply(lambda x: x + x.min()).sort_index(),
         )
 
     def test_transform(self):
@@ -367,132 +258,6 @@ class OpsOnDiffFramesGroupByTestsMixin:
         self.assert_eq(
             pdf.groupby([pkey, "b"]).head(2).sort_index(),
             psdf.groupby([kkey, "b"]).head(2).sort_index(),
-        )
-
-    def test_cumcount(self):
-        pdf = pd.DataFrame(
-            {
-                "a": [1, 2, 3, 4, 5, 6] * 3,
-                "b": [1, 1, 2, 3, 5, 8] * 3,
-                "c": [1, 4, 9, 16, 25, 36] * 3,
-            },
-        )
-        pkey = pd.Series([1, 1, 2, 3, 5, 8] * 3)
-        psdf = ps.from_pandas(pdf)
-        kkey = ps.from_pandas(pkey)
-
-        for ascending in [True, False]:
-            self.assert_eq(
-                psdf.groupby(kkey).cumcount(ascending=ascending).sort_index(),
-                pdf.groupby(pkey).cumcount(ascending=ascending).sort_index(),
-            )
-            self.assert_eq(
-                psdf.groupby(kkey)["a"].cumcount(ascending=ascending).sort_index(),
-                pdf.groupby(pkey)["a"].cumcount(ascending=ascending).sort_index(),
-            )
-            self.assert_eq(
-                psdf.groupby(kkey)[["a"]].cumcount(ascending=ascending).sort_index(),
-                pdf.groupby(pkey)[["a"]].cumcount(ascending=ascending).sort_index(),
-            )
-
-    def test_cummin(self):
-        pdf = pd.DataFrame(
-            {
-                "a": [1, 2, 3, 4, 5, 6] * 3,
-                "b": [1, 1, 2, 3, 5, 8] * 3,
-                "c": [1, 4, 9, 16, 25, 36] * 3,
-            },
-        )
-        pkey = pd.Series([1, 1, 2, 3, 5, 8] * 3)
-        psdf = ps.from_pandas(pdf)
-        kkey = ps.from_pandas(pkey)
-
-        self.assert_eq(
-            psdf.groupby(kkey).cummin().sort_index(), pdf.groupby(pkey).cummin().sort_index()
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)["a"].cummin().sort_index(),
-            pdf.groupby(pkey)["a"].cummin().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)[["a"]].cummin().sort_index(),
-            pdf.groupby(pkey)[["a"]].cummin().sort_index(),
-        )
-
-    def test_cummax(self):
-        pdf = pd.DataFrame(
-            {
-                "a": [1, 2, 3, 4, 5, 6] * 3,
-                "b": [1, 1, 2, 3, 5, 8] * 3,
-                "c": [1, 4, 9, 16, 25, 36] * 3,
-            },
-        )
-        pkey = pd.Series([1, 1, 2, 3, 5, 8] * 3)
-        psdf = ps.from_pandas(pdf)
-        kkey = ps.from_pandas(pkey)
-
-        self.assert_eq(
-            psdf.groupby(kkey).cummax().sort_index(), pdf.groupby(pkey).cummax().sort_index()
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)["a"].cummax().sort_index(),
-            pdf.groupby(pkey)["a"].cummax().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)[["a"]].cummax().sort_index(),
-            pdf.groupby(pkey)[["a"]].cummax().sort_index(),
-        )
-
-    def test_cumsum(self):
-        pdf = pd.DataFrame(
-            {
-                "a": [1, 2, 3, 4, 5, 6] * 3,
-                "b": [1, 1, 2, 3, 5, 8] * 3,
-                "c": [1, 4, 9, 16, 25, 36] * 3,
-            },
-        )
-        pkey = pd.Series([1, 1, 2, 3, 5, 8] * 3)
-        psdf = ps.from_pandas(pdf)
-        kkey = ps.from_pandas(pkey)
-
-        self.assert_eq(
-            psdf.groupby(kkey).cumsum().sort_index(), pdf.groupby(pkey).cumsum().sort_index()
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)["a"].cumsum().sort_index(),
-            pdf.groupby(pkey)["a"].cumsum().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)[["a"]].cumsum().sort_index(),
-            pdf.groupby(pkey)[["a"]].cumsum().sort_index(),
-        )
-
-    def test_cumprod(self):
-        pdf = pd.DataFrame(
-            {
-                "a": [1, 2, 3, 4, 5, 6] * 3,
-                "b": [1, 1, 2, 3, 5, 8] * 3,
-                "c": [1, 4, 9, 16, 25, 36] * 3,
-            },
-        )
-        pkey = pd.Series([1, 1, 2, 3, 5, 8] * 3)
-        psdf = ps.from_pandas(pdf)
-        kkey = ps.from_pandas(pkey)
-
-        self.assert_eq(
-            psdf.groupby(kkey).cumprod().sort_index(),
-            pdf.groupby(pkey).cumprod().sort_index(),
-            almost=True,
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)["a"].cumprod().sort_index(),
-            pdf.groupby(pkey)["a"].cumprod().sort_index(),
-            almost=True,
-        )
-        self.assert_eq(
-            psdf.groupby(kkey)[["a"]].cumprod().sort_index(),
-            pdf.groupby(pkey)[["a"]].cumprod().sort_index(),
-            almost=True,
         )
 
     def test_diff(self):
